requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
apturl==0.5.2
asciinema==2.0.2
async-timeout==4.0.2
attrs==22.2.0
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decouple==0.0.7
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
frozenlist==1.3.3
future==0.18.2
httplib2==0.14.0
idna==2.8
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.1.2
monotonic==1.5
multidict==6.0.4
netifaces==0.10.4
numpy==1.17.4
oauthlib==3.1.0
olefile==0.46
openai==0.26.5
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyRFC3339==1.1
python-apt==2.0.1
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.7
python-dotenv==0.21.1
pythondialog==3.5.2
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
systemd-python==234
tqdm==4.64.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
vboxapi==1.0
wadllib==1.3.3
xkit==0.0.0
yarl==1.8.2
